blog-x/requirements.txt
# COYASS Auto-Posting System
# Core
python-dotenv==1.1.0
pyyaml==6.0.2
apscheduler==3.11.0

# AI Content Generation
openai==1.68.0
google-generativeai==0.8.5

# X (Twitter) API
tweepy==4.15.0

# Note.com Browser Automation
playwright==1.50.0

# Database
sqlalchemy==2.0.38

# Dashboard
flask==3.1.0

# Image handling
Pillow==11.1.0

# Utilities
requests==2.32.3
jinja2==3.1.5
rich==13.9.4
